Decompose machine semantic internals
