+++ src/internal/machine/planner.ts
@@ -9,7 +9,7 @@ import * as Effect from "effect/Effect"
 import * as Option from "effect/Option"
 import type * as Schema from "effect/Schema"
 import type { Command, Enqueue, InitialEvent as MachineInitialEvent, Machine, Runtime } from "../../Machine.js"
-import { InfiniteTransitionError, MachineSchemaDecodeError, StartupError } from "./errors.js"
+import { getTargetBuilder, makeCollector, type RuntimeCommand } from "./command.js"
 import {
   type ActiveConfiguration,
   captureHistory,
@@ -17,34 +17,19 @@ import {
   completeConfigurationEffect,
   completeConfigurationSync,
   configurationFromHistoryRecord,
-  decodeEmit,
-  decodeEmitSync,
-  decodeEvent,
-  decodeEventSync,
-  decodeInput,
-  decodeInputSync,
-  decodeStateValue,
-  decodeStateValueSync,
   getActiveLeafPathFrom,
   getActiveLeafPaths,
   getActiveValue,
   getHistoryRecord,
   getInitialEntryPaths,
   getLeafPath,
-  getNode,
   getParentValue,
   getParentValues,
   getPathToRoot,
   getRootPath,
   isActiveFinalConfiguration,
-  isChoiceTarget,
   isDescendantOf,
-  isHistoryTarget,
   isPathInSubtree,
-  isSnapshot,
-  isTarget,
-  makeChoiceTarget,
-  makeTarget,
   normalizeConfiguration,
   normalizeConfigurationEffect,
   normalizeConfigurationSync,
@@ -53,93 +38,30 @@ import {
   pathDepth,
   snapshotFromConfiguration,
   snapshotFromConfigurationAtPath,
-  TargetSnapshotTypeId,
   validateInitialConfiguration
-} from "./model.js"
-import type { ProcessScope } from "./runtime.js"
+} from "./configuration.js"
+import { InfiniteTransitionError, MachineSchemaDecodeError, StartupError } from "./errors.js"
+import {
+  decodeEmit,
+  decodeEmitSync,
+  decodeEvent,
+  decodeEventSync,
+  decodeInput,
+  decodeInputSync,
+  decodeStateValue,
+  decodeStateValueSync
+} from "./protocol.js"
 import { InitialEventTypeId } from "./symbols.js"
-
-export type RuntimeCommand = Command
-
-interface Collected<Event> {
-  readonly enqueue: Enqueue<Event, unknown>
-  readonly commands: Array<RuntimeCommand>
-  readonly raisedEvents: Array<Event>
-  readonly emittedEvents: Array<unknown>
-}
-
-const targetBuilderCache = new WeakMap<object, Map<string, unknown>>()
-
-const getTargetBuilder = (machine: Machine.Any, path: string): any => {
-  let byPath = targetBuilderCache.get(machine)
-  if (byPath === undefined) {
-    byPath = new Map()
-    targetBuilderCache.set(machine, byPath)
-  }
-  if (byPath.has(path)) {
-    return byPath.get(path)
-  }
-  const builder = machine.makeTargetBuilder(path as any)
-  byPath.set(path, builder)
-  return builder
-}
-
-const makeCollector = <Event>(machine: Machine.Any): Collected<Event> => {
-  const commands: Array<RuntimeCommand> = []
-  const raisedEvents: Array<Event> = []
-  const emittedEvents: Array<unknown> = []
-  return {
-    commands,
-    raisedEvents,
-    emittedEvents,
-    enqueue: {
-      raise: (event) => {
-        raisedEvents.push(decodeEventSync(machine, event) as Event)
-      },
-      emit: (event) => {
-        emittedEvents.push(decodeEmitSync(machine, event))
-      },
-      sendTo: (child: unknown, event: unknown) => {
-        commands.push({ _tag: "SendTo", child: child as any, event })
-      },
-      stop: (child: unknown) => {
-        commands.push({ _tag: "Stop", child: child as any })
-      }
-    }
-  }
-}
-
-export const makeLiveRuntime = <Events, Emits>(
-  machine: Machine.Any,
-  scope: ProcessScope<Events>
-): Runtime<Events, Emits> => ({
-  raise: (event) =>
-    decodeEvent(machine, event).pipe(
-      Effect.flatMap((event) => scope.self.send(event as Events))
-    ),
-  sendParent: (event) =>
-    decodeEmit(machine, event).pipe(
-      Effect.flatMap((event) => scope.sendParent(event))
-    )
-})
-
-export const runCommands = <Event>(
-  commands: Iterable<RuntimeCommand>,
-  scope: ProcessScope<Event>
-) =>
-  Effect.forEach(commands, (command) =>
-    command._tag === "SendTo"
-      ? scope.sendTo(command.child as never, command.event)
-      : scope.stopChild(command.child as never), { discard: true })
-
-export const runEmittedEvents = <Events, Emits>(
-  events: Iterable<Emits>,
-  runtime: Runtime<Events, Emits>
-) =>
-  Effect.all(
-    Array.from(events, (event) => runtime.sendParent(event)),
-    { discard: true }
-  )
+import {
+  getNode,
+  isChoiceTarget,
+  isHistoryTarget,
+  isSnapshot,
+  isTarget,
+  makeChoiceTarget,
+  makeTarget,
+  TargetSnapshotTypeId
+} from "./topology.js"
 
 export type MicrostepPlan<State, Event, E, R> = {
   readonly next: State
@@ -177,7 +99,7 @@ export type MacrostepPlan<State, Event, E, R, Output> =
     }
   )
 
-type TransitionHandler<States extends Machine.StateSchemas, E, R, Context> = (
+export type TransitionHandler<States extends Machine.StateSchemas, E, R, Context> = (
   context: Context,
   enqueue: Enqueue<any, any>
 ) => Machine.HandlerResult<States, E, R>
@@ -190,13 +112,13 @@ type EventTransition<States extends Machine.StateSchemas, E, R, Context> =
     readonly transition: TransitionHandler<States, E, R, Context>
   }
 
-type MicrostepTransition<States extends Machine.StateSchemas, E, R, Context> = {
+export type MicrostepTransition<States extends Machine.StateSchemas, E, R, Context> = {
   readonly reenter: boolean
   readonly targets: ReadonlyArray<string> | undefined
   readonly transition: TransitionHandler<States, E, R, Context>
 }
 
-const normalizeTransition = <States extends Machine.StateSchemas, E, R, Context>(
+export const normalizeTransition = <States extends Machine.StateSchemas, E, R, Context>(
   transition: EventTransition<States, E, R, Context> | undefined
 ): MicrostepTransition<States, E, R, Context> | undefined => {
   if (transition === undefined) {
@@ -462,7 +384,7 @@ const resolveHistoryTarget = (
   }
 }
 
-type SelectedTransition<States extends Machine.StateSchemas, E, R, Context> = {
+export type SelectedTransition<States extends Machine.StateSchemas, E, R, Context> = {
   readonly sourcePath: string
   readonly leafPath: string
   readonly trigger: Machine.TransitionTrigger
@@ -470,7 +392,7 @@ type SelectedTransition<States extends Machine.StateSchemas, E, R, Context> = {
   readonly context: Context
 }
 
-type EvaluatedTransition<States extends Machine.StateSchemas, Event, E, R, Context> = {
+export type EvaluatedTransition<States extends Machine.StateSchemas, Event, E, R, Context> = {
   readonly selection: SelectedTransition<States, E, R, Context>
   readonly unresolvedTarget:
     | Machine.Snapshot<States>
@@ -507,7 +429,7 @@ const getCandidatePaths = (machine: Machine.Any, configuration: ActiveConfigurat
 const getLeafCandidatePaths = (machine: Machine.Any, leaf: string): ReadonlyArray<string> =>
   [...getPathToRoot(machine, leaf)].reverse()
 
-const getLeastCommonAncestor = (
+export const getLeastCommonAncestor = (
   machine: Machine.Any,
   left: string,
   right: string
@@ -525,7 +447,7 @@ const getLeastCommonAncestor = (
   return ancestor
 }
 
-const broadenTransitionBoundary = (
+export const broadenTransitionBoundary = (
   naturalBoundary: string | undefined,
   reentryBoundary: string | undefined
 ): string | undefined => {
@@ -535,7 +457,7 @@ const broadenTransitionBoundary = (
     : naturalBoundary
 }
 
-const getExitPaths = (
+export const getExitPaths = (
   machine: Machine.Any,
   configuration: ActiveConfiguration,
   boundary: string | undefined
@@ -546,7 +468,7 @@ const getExitPaths = (
       .filter((path) => boundary === undefined || isDescendantOf(path, boundary))
   )
 
-const getEntryPaths = (
+export const getEntryPaths = (
   machine: Machine.Any,
   configuration: ActiveConfiguration,
   boundary: string | undefined
@@ -878,7 +800,7 @@ const selectEventTransitions = <
   return selected
 }
 
-const getTargetNodePath = <const States extends Machine.StateSchemas>(
+export const getTargetNodePath = <const States extends Machine.StateSchemas>(
   target:
     | Machine.Snapshot<States>
     | Machine.Target<States, Machine.StateIdentifier<States>>
@@ -897,7 +819,7 @@ const getTargetNodePath = <const States extends Machine.StateSchemas>(
   throw new Error("Machine expected transition target to be a snapshot or target builder result")
 }
 
-const validateDeclaredTransitionTarget = (
+export const validateDeclaredTransitionTarget = (
   sourcePath: string,
   trigger: Machine.TransitionTrigger,
   declaredTargets: ReadonlyArray<string> | undefined,
@@ -928,7 +850,7 @@ const hasPathIntersection = (left: ReadonlyArray<string>, right: ReadonlyArray<s
   return false
 }
 
-const sortEvaluatedTransitions = <
+export const sortEvaluatedTransitions = <
   const States extends Machine.StateSchemas,
   Event,
   E,
@@ -955,7 +877,7 @@ const removePreemptedAncestorSelections = <
     )
   )
 
-const removeConflictingTransitions = <
+export const removeConflictingTransitions = <
   const States extends Machine.StateSchemas,
   Event,
   E,
@@ -1343,7 +1265,7 @@ const collectEvaluatedTransition = <
   } as EvaluatedTransition<States, Event, E, R, Context>
 }
 
-const MaxMacrostepIterations = 1000
+export const MaxMacrostepIterations = 1000
 export const InitialEvent: MachineInitialEvent = { _tag: InitialEventTypeId }
 
 const catchStartup = <A, E, R>(
@@ -1944,884 +1866,6 @@ const macrostepConfiguration = <
   const emittedEvents = [...step.emittedEvents]
   const microsteps = [step]
   return settle(machine, step.next, decodedEvent, commands, raisedEvents, emittedEvents, microsteps)
-}
-
-interface IndexedExecutionDescriptor {
-  readonly flat: boolean
-  readonly nodes: ReadonlyArray<Machine.StateNode>
-  readonly indexByPath: ReadonlyMap<string, number>
-  readonly parentIndices: ReadonlyArray<number>
-  readonly childIndices: ReadonlyArray<ReadonlyArray<number>>
-  readonly ancestorIndices: ReadonlyArray<ReadonlyArray<number>>
-  readonly rootIndices: ReadonlyArray<number>
-  readonly leafIndices: ReadonlyArray<number>
-  readonly finalIndices: ReadonlyArray<number>
-  readonly dispatchByLeaf: ReadonlyMap<
-    number,
-    ReadonlyMap<PropertyKey, {
-      readonly sourceIndex: number
-      readonly transition: MicrostepTransition<any, any, any, any>
-    }>
-  >
-}
-
-const compileIndexedExecutionDescriptor = (
-  machine: Machine.Any
-): IndexedExecutionDescriptor | undefined => {
-  const nodes: Array<Machine.StateNode> = []
-  const leafPaths: Array<string> = []
-  const finalPaths: Array<string> = []
-  const transitionsByPath = new Map<
-    PropertyKey,
-    ReadonlyMap<PropertyKey, MicrostepTransition<any, any, any, any>>
-  >()
-
-  for (const node of machine.stateNodes.byPath.values() as Iterable<Machine.StateNode>) {
-    nodes.push(node)
-    if (node.type === "choice" || node.type === "history") {
-      return undefined
-    }
-    if (node.type === "atomic" || node.type === "final") {
-      leafPaths.push(node.path)
-    }
-    if (node.type === "final") {
-      finalPaths.push(node.path)
-    }
-
-    const config = machine.handlers[node.path] as Machine.AnyStateConfig | undefined
-    if (
-      config?.entry !== undefined || config?.exit !== undefined || config?.always !== undefined ||
-      config?.onDone !== undefined || (config as any)?.choice !== undefined ||
-      (config as any)?.history !== undefined
-    ) {
-      return undefined
-    }
-    if (config?.on === undefined) {
-      continue
-    }
-    const byEvent = new Map<PropertyKey, MicrostepTransition<any, any, any, any>>()
-    for (const tag of Reflect.ownKeys(config.on)) {
-      const transition = normalizeTransition(config.on[tag])
-      if (transition !== undefined) {
-        byEvent.set(tag, transition)
-      }
-    }
-    if (byEvent.size > 0) {
-      transitionsByPath.set(node.path, byEvent)
-    }
-  }
-
-  leafPaths.sort((left, right) => compareDocumentOrder(machine, left, right))
-  finalPaths.sort((left, right) => compareDocumentOrder(machine, left, right))
-  nodes.sort((left, right) => left.order - right.order)
-  const indexByPath = new Map(nodes.map((node, index) => [node.path, index]))
-  const indexOf = (path: string): number => {
-    const index = indexByPath.get(path)
-    if (index === undefined) {
-      throw new Error(`Machine expected compiled state path "${path}"`)
-    }
-    return index
-  }
-  const leafIndices = leafPaths.map(indexOf)
-  const parentIndices = nodes.map((node) => node.parent === undefined ? -1 : indexOf(node.parent))
-  const childIndices = nodes.map((node) => node.children.map(indexOf))
-  const ancestorIndices = nodes.map((node) => getPathToRoot(machine, node.path).slice(0, -1).map(indexOf))
-  const transitionsByIndex = nodes.map((node) => transitionsByPath.get(node.path))
-  const dispatchByLeaf = new Map<
-    number,
-    ReadonlyMap<PropertyKey, {
-      readonly sourceIndex: number
-      readonly transition: MicrostepTransition<any, any, any, any>
-    }>
-  >()
-  for (const leafIndex of leafIndices) {
-    const dispatch = new Map<PropertyKey, {
-      readonly sourceIndex: number
-      readonly transition: MicrostepTransition<any, any, any, any>
-    }>()
-    const candidates = [leafIndex, ...ancestorIndices[leafIndex]!.slice().reverse()]
-    for (const sourceIndex of candidates) {
-      for (const [tag, transition] of transitionsByIndex[sourceIndex] ?? []) {
-        if (!dispatch.has(tag)) dispatch.set(tag, { sourceIndex, transition })
-      }
-    }
-    dispatchByLeaf.set(leafIndex, dispatch)
-  }
-  return {
-    flat: nodes.every((node) => node.parent === undefined && (node.type === "atomic" || node.type === "final")),
-    nodes,
-    indexByPath,
-    parentIndices,
-    childIndices,
-    ancestorIndices,
-    rootIndices: nodes.flatMap((node, index) => node.parent === undefined ? [index] : []),
-    leafIndices,
-    finalIndices: finalPaths.map(indexOf),
-    dispatchByLeaf
-  }
-}
-
-interface IndexedConfiguration {
-  readonly active: Uint8Array
-  readonly activeLeaves: ReadonlyArray<number>
-  // The compiled process fiber owns this slot table. Flat same-state updates
-  // may replace one value in place after eagerly detaching the handler's
-  // public snapshot; hierarchical microsteps retain immutable copies so
-  // simultaneous transition contexts continue to share their starting state.
-  readonly values: Array<unknown>
-  readonly completed: Uint8Array
-  readonly outputs: ReadonlyArray<unknown>
-  readonly completedOrder: ReadonlyArray<number>
-}
-
-const indexedConfigurationFromActive = (
-  descriptor: IndexedExecutionDescriptor,
-  configuration: ActiveConfiguration
-): IndexedConfiguration => {
-  const active = new Uint8Array(descriptor.nodes.length)
-  const values: Array<unknown> = new Array(descriptor.nodes.length)
-  const completed = new Uint8Array(descriptor.nodes.length)
-  const outputs: Array<unknown> = new Array(descriptor.nodes.length)
-  const completedOrder: Array<number> = []
-  for (const path of configuration.active) {
-    const index = descriptor.indexByPath.get(path)
-    if (index === undefined) throw new Error(`Machine expected indexed active path "${path}"`)
-    active[index] = 1
-    values[index] = configuration.values.get(path)
-  }
-  for (const [path, output] of configuration.outputs) {
-    const index = descriptor.indexByPath.get(path)
-    if (index === undefined) throw new Error(`Machine expected indexed completed path "${path}"`)
-    completed[index] = 1
-    outputs[index] = output
-    completedOrder.push(index)
-  }
-  return {
-    active,
-    activeLeaves: descriptor.leafIndices.filter((index) => active[index] === 1),
-    values,
-    completed,
-    outputs,
-    completedOrder
-  }
-}
-
-const activeConfigurationFromIndexed = (
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration
-): ActiveConfiguration => {
-  const active = new Set<string>()
-  const values = new Map<string, unknown>()
-  const outputs = new Map<string, unknown>()
-  for (let index = 0; index < descriptor.nodes.length; index++) {
-    if (configuration.active[index] !== 1) continue
-    const path = descriptor.nodes[index]!.path
-    active.add(path)
-    values.set(path, configuration.values[index])
-  }
-  for (const index of configuration.completedOrder) {
-    if (configuration.completed[index] === 1) {
-      outputs.set(descriptor.nodes[index]!.path, configuration.outputs[index])
-    }
-  }
-  return { active, values, outputs, history: new Map() }
-}
-
-const snapshotFromIndexedPath = (
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration,
-  index: number
-): Machine.AtomicSnapshot<string, unknown> => {
-  const node = descriptor.nodes[index]!
-  const snapshot: Record<string, unknown> = {
-    path: node.path,
-    value: configuration.values[index]
-  }
-  if (node.type === "compound") {
-    const childIndex = descriptor.childIndices[index]!.find((childIndex) => configuration.active[childIndex] === 1)
-    if (childIndex === undefined) {
-      throw new Error(`Machine expected indexed compound state "${node.path}" to have an active child`)
-    }
-    snapshot.state = snapshotFromIndexedPath(descriptor, configuration, childIndex)
-  } else if (node.type === "parallel") {
-    const states: Record<string, unknown> = {}
-    for (const childIndex of descriptor.childIndices[index]!) {
-      if (configuration.active[childIndex] !== 1) {
-        throw new Error(
-          `Machine expected indexed parallel state "${node.path}" to have active region "${
-            descriptor.nodes[childIndex]!.path
-          }"`
-        )
-      }
-      states[descriptor.nodes[childIndex]!.key] = snapshotFromIndexedPath(descriptor, configuration, childIndex)
-    }
-    snapshot.states = states
-  }
-  return snapshot as unknown as Machine.AtomicSnapshot<string, unknown>
-}
-
-const snapshotFromIndexed = (
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration
-): Machine.Snapshot<any> => {
-  const rootIndex = descriptor.rootIndices.find((index) => configuration.active[index] === 1)
-  if (rootIndex === undefined) throw new Error("Machine expected an active indexed root state")
-  const snapshot = snapshotFromIndexedPath(descriptor, configuration, rootIndex) as Machine.Snapshot<any>
-  if (configuration.completedOrder.length > 0) {
-    ;(snapshot as Machine.AtomicSnapshot<string, unknown> & {
-      completed: ReadonlyArray<Machine.SnapshotCompletion>
-    }).completed = configuration.completedOrder.map((index) => ({
-      path: descriptor.nodes[index]!.path,
-      output: configuration.outputs[index]
-    }))
-  }
-  return snapshot
-}
-
-const makeIndexedTransitionContext = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration,
-  sourceIndex: number,
-  event: any
-): any => {
-  const source = descriptor.nodes[sourceIndex]!
-  const parentIndex = descriptor.parentIndices[sourceIndex]!
-  const parents: Record<string, unknown> = {}
-  for (const ancestorIndex of descriptor.ancestorIndices[sourceIndex]!) {
-    parents[descriptor.nodes[ancestorIndex]!.path] = configuration.values[ancestorIndex]
-  }
-  return {
-    state: configuration.values[sourceIndex],
-    parent: parentIndex < 0 ? undefined : configuration.values[parentIndex],
-    parents,
-    event,
-    snapshot: snapshotFromIndexed(descriptor, configuration),
-    target: getTargetBuilder(machine, source.path)
-  }
-}
-
-type IndexedSelectedTransition = SelectedTransition<any, any, any, any> & {
-  readonly sourceIndex: number
-  readonly leafIndex: number
-}
-
-type IndexedEvaluatedTransition =
-  & Omit<
-    EvaluatedTransition<any, any, any, any, any>,
-    "selection"
-  >
-  & {
-    readonly selection: IndexedSelectedTransition
-    readonly next: IndexedConfiguration
-  }
-
-const emptyCompiledValues: ReadonlyArray<never> = []
-
-const collectIndexedTransition = (
-  machine: Machine.Any,
-  transition: TransitionHandler<any, any, any, any>,
-  context: any
-) => {
-  let commands: Array<RuntimeCommand> | undefined
-  let raisedEvents: Array<any> | undefined
-  let emittedEvents: Array<unknown> | undefined
-  const state = transition(context, {
-    raise: (event: unknown) => {
-      ;(raisedEvents ??= []).push(decodeEventSync(machine, event))
-    },
-    emit: (event: unknown) => {
-      ;(emittedEvents ??= []).push(decodeEmitSync(machine, event))
-    },
-    sendTo: (child: unknown, event: unknown) => {
-      ;(commands ??= []).push({ _tag: "SendTo", child: child as any, event })
-    },
-    stop: (child: unknown) => {
-      ;(commands ??= []).push({ _tag: "Stop", child: child as any })
-    }
-  })
-  return {
-    state,
-    commands: commands ?? emptyCompiledValues,
-    raisedEvents: raisedEvents ?? emptyCompiledValues,
-    emittedEvents: emittedEvents ?? emptyCompiledValues
-  }
-}
-
-const selectIndexedEventTransitions = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration,
-  event: any
-): ReadonlyArray<IndexedSelectedTransition> => {
-  const selected: Array<IndexedSelectedTransition> = []
-  for (const leafIndex of configuration.activeLeaves) {
-    const dispatched = descriptor.dispatchByLeaf.get(leafIndex)!.get(event._tag)
-    if (dispatched !== undefined) {
-      const { sourceIndex, transition } = dispatched
-      if (!selected.some((selection) => selection.sourceIndex === sourceIndex)) {
-        const sourcePath = descriptor.nodes[sourceIndex]!.path
-        selected.push({
-          sourceIndex,
-          leafIndex,
-          sourcePath,
-          leafPath: descriptor.nodes[leafIndex]!.path,
-          trigger: { type: "event", event: event._tag },
-          transition,
-          context: makeIndexedTransitionContext(
-            machine,
-            descriptor,
-            configuration,
-            sourceIndex,
-            event
-          )
-        })
-      }
-    }
-  }
-  return selected
-}
-
-const hasSameIndexedActive = (left: IndexedConfiguration, right: IndexedConfiguration): boolean => {
-  if (left.active === right.active) return true
-  for (let index = 0; index < left.active.length; index++) {
-    if (left.active[index] !== right.active[index]) return false
-  }
-  return true
-}
-
-const normalizeIndexedTargetConfigurationSync = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  current: IndexedConfiguration,
-  target: Machine.Target<any, any> | Machine.Snapshot<any>,
-  activeLeafIndex: number
-): IndexedConfiguration => {
-  const targetIndex = isTarget(target) ? descriptor.indexByPath.get(String(target.path)) : undefined
-  if (
-    targetIndex === activeLeafIndex && current.active[activeLeafIndex] === 1 && isTarget(target) &&
-    target[TargetSnapshotTypeId] === undefined && target.values === undefined && current.completedOrder.length === 0
-  ) {
-    const values = current.values.slice()
-    values[activeLeafIndex] = decodeStateValueSync(
-      machine,
-      descriptor.nodes[activeLeafIndex]!,
-      target.value
-    )
-    return { ...current, values }
-  }
-  return indexedConfigurationFromActive(
-    descriptor,
-    normalizeTargetConfigurationSync(
-      machine,
-      activeConfigurationFromIndexed(descriptor, current),
-      target
-    )
-  )
-}
-
-const collectIndexedEvaluatedTransition = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  state: IndexedConfiguration,
-  selection: IndexedSelectedTransition
-): IndexedEvaluatedTransition => {
-  const transitionResult = collectIndexedTransition(machine, selection.transition.transition, selection.context)
-  const target = transitionResult.state
-  validateDeclaredTransitionTarget(
-    selection.sourcePath,
-    selection.trigger,
-    selection.transition.targets,
-    target
-  )
-  if (target !== undefined && !isTarget(target) && !isSnapshot(target)) {
-    throw new Error("Machine expected indexed transition target to be a snapshot or target builder result")
-  }
-  const next = target === undefined
-    ? state
-    : normalizeIndexedTargetConfigurationSync(machine, descriptor, state, target as any, selection.leafIndex)
-  const changed = selection.transition.reenter || !hasSameIndexedActive(state, next)
-  if (!changed) {
-    return {
-      selection,
-      unresolvedTarget: target as any,
-      target: target as any,
-      next,
-      commands: transitionResult.commands,
-      raisedEvents: transitionResult.raisedEvents,
-      emittedEvents: transitionResult.emittedEvents,
-      changed: false,
-      exitPaths: [],
-      entryPaths: [],
-      choiceTransitions: []
-    }
-  }
-
-  const targetPath = target === undefined ? undefined : getTargetNodePath(target as any)
-  const naturalBoundary = targetPath === undefined
-    ? descriptor.nodes[selection.sourceIndex]!.parent
-    : getLeastCommonAncestor(machine, selection.leafPath, targetPath)
-  const reentryBoundary = descriptor.nodes[selection.sourceIndex]!.parent
-  const boundary = selection.transition.reenter
-    ? broadenTransitionBoundary(naturalBoundary, reentryBoundary)
-    : naturalBoundary
-  return {
-    selection,
-    unresolvedTarget: target as any,
-    target: target as any,
-    next,
-    commands: transitionResult.commands,
-    raisedEvents: transitionResult.raisedEvents,
-    emittedEvents: transitionResult.emittedEvents,
-    changed: true,
-    exitPaths: getExitPaths(machine, activeConfigurationFromIndexed(descriptor, state), boundary),
-    entryPaths: getEntryPaths(machine, activeConfigurationFromIndexed(descriptor, next), boundary),
-    choiceTransitions: []
-  }
-}
-
-const indexedMicrostep = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  state: IndexedConfiguration,
-  event: any,
-  selections: ReadonlyArray<IndexedSelectedTransition>
-): MicrostepPlan<IndexedConfiguration, any, any, any> => {
-  if (selections.length === 1) {
-    const transition = collectIndexedEvaluatedTransition(machine, descriptor, state, selections[0]!)
-    return {
-      next: transition.next,
-      event,
-      transitions: [],
-      commands: transition.commands,
-      raisedEvents: transition.raisedEvents,
-      emittedEvents: transition.emittedEvents,
-      exitPaths: transition.exitPaths,
-      entryPaths: transition.entryPaths,
-      changed: transition.changed
-    }
-  }
-  const activeSelections = selections.filter((selection) =>
-    !selections.some((other) =>
-      other.sourceIndex !== selection.sourceIndex &&
-      isDescendantOf(other.sourcePath, selection.sourcePath)
-    )
-  )
-  const evaluated = activeSelections.map((selection) =>
-    collectIndexedEvaluatedTransition(machine, descriptor, state, selection)
-  )
-  const transitions = sortEvaluatedTransitions(
-    machine,
-    removeConflictingTransitions(machine, evaluated as any)
-  ) as ReadonlyArray<IndexedEvaluatedTransition>
-
-  let next = state
-  if (transitions.length === 1) {
-    next = transitions[0]!.next
-  } else {
-    const applicationOrder = [
-      ...transitions.filter((transition) => !transition.changed),
-      ...transitions.filter((transition) => transition.changed)
-    ]
-    for (const transition of applicationOrder) {
-      if (transition.target !== undefined) {
-        next = normalizeIndexedTargetConfigurationSync(
-          machine,
-          descriptor,
-          next,
-          transition.target,
-          transition.selection.leafIndex
-        )
-      }
-    }
-  }
-
-  const commands = transitions.flatMap((transition) => transition.commands)
-  const raisedEvents = transitions.flatMap((transition) => transition.raisedEvents)
-  const emittedEvents = transitions.flatMap((transition) => transition.emittedEvents)
-  const changed = transitions.some((transition) => transition.changed)
-  return {
-    next,
-    event,
-    transitions: [],
-    commands,
-    raisedEvents,
-    emittedEvents,
-    exitPaths: changed ? sortExitPaths(machine, transitions.flatMap((transition) => transition.exitPaths)) : [],
-    entryPaths: changed ? sortEntryPaths(machine, transitions.flatMap((transition) => transition.entryPaths)) : [],
-    changed
-  }
-}
-
-const planIndexedFlatConfiguration = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration,
-  decoded: { readonly _tag: PropertyKey }
-): MacrostepPlan<IndexedConfiguration, any, any, any, any> => {
-  let current = configuration
-  let event: any = decoded
-  let commands: Array<RuntimeCommand> | undefined
-  let raisedEvents: Array<any> | undefined
-  let emittedEvents: Array<unknown> | undefined
-  let microsteps: Array<MicrostepPlan<IndexedConfiguration, any, any, any>> | undefined
-  let raisedIndex = 0
-  let iterations = 0
-
-  while (true) {
-    iterations += 1
-    if (iterations > MaxMacrostepIterations) {
-      throw new InfiniteTransitionError({
-        machineId: machine.id,
-        state: descriptor.nodes[current.activeLeaves[0]!]!.path,
-        maxIterations: MaxMacrostepIterations
-      })
-    }
-
-    const sourceIndex = current.activeLeaves[0]
-    if (sourceIndex === undefined) {
-      throw new Error("Machine expected an active indexed root state")
-    }
-    if (descriptor.nodes[sourceIndex]!.type === "final") {
-      const completed = completeConfigurationSync(
-        machine,
-        activeConfigurationFromIndexed(descriptor, current),
-        event
-      ).configuration
-      const root = getRootPath(machine, completed)
-      if (!completed.outputs.has(root)) {
-        throw new Error("Machine reached a terminal indexed configuration without a completed root output")
-      }
-      return {
-        next: indexedConfigurationFromActive(descriptor, completed),
-        commands: commands ?? emptyCompiledValues,
-        emittedEvents: emittedEvents ?? emptyCompiledValues,
-        microsteps: microsteps ?? emptyCompiledValues,
-        done: true,
-        output: completed.outputs.get(root)
-      }
-    }
-
-    const sourcePath = descriptor.nodes[sourceIndex]!.path
-    const transition = normalizeTransition(machine.handlers[sourcePath]?.on?.[event._tag])
-    if (transition !== undefined) {
-      const transitionResult = collectIndexedTransition(
-        machine,
-        transition.transition,
-        {
-          state: current.values[sourceIndex],
-          parent: undefined,
-          parents: {},
-          event,
-          snapshot: snapshotFromIndexed(descriptor, current),
-          target: getTargetBuilder(machine, sourcePath)
-        }
-      )
-      const target = transitionResult.state
-      validateDeclaredTransitionTarget(
-        sourcePath,
-        { type: "event", event: event._tag },
-        transition.targets,
-        target
-      )
-      if (target !== undefined && !isTarget(target) && !isSnapshot(target)) {
-        throw new Error("Machine expected indexed transition target to be a snapshot or target builder result")
-      }
-
-      let next = current
-      if (target !== undefined) {
-        const targetIndex = descriptor.indexByPath.get(String(target.path))
-        const isSimpleTarget = isTarget(target)
-          ? target[TargetSnapshotTypeId] === undefined && target.values === undefined
-          : !("state" in target) && !("states" in target) && !("completed" in target) && !("history" in target)
-        if (
-          targetIndex === sourceIndex && isSimpleTarget && current.completedOrder.length === 0
-        ) {
-          current.values[sourceIndex] = decodeStateValueSync(
-            machine,
-            descriptor.nodes[sourceIndex]!,
-            target.value
-          )
-        } else {
-          next = normalizeIndexedTargetConfigurationSync(machine, descriptor, current, target as any, sourceIndex)
-        }
-      }
-      const changed = transition.reenter || !hasSameIndexedActive(current, next)
-      const nextIndex = next.activeLeaves[0]
-      if (nextIndex === undefined) {
-        throw new Error("Machine expected an active indexed transition target")
-      }
-      const step: MicrostepPlan<IndexedConfiguration, any, any, any> = {
-        next,
-        event,
-        transitions: emptyCompiledValues,
-        commands: transitionResult.commands,
-        raisedEvents: transitionResult.raisedEvents,
-        emittedEvents: transitionResult.emittedEvents,
-        exitPaths: changed ? [sourcePath] : emptyCompiledValues,
-        entryPaths: changed ? [descriptor.nodes[nextIndex]!.path] : emptyCompiledValues,
-        changed
-      }
-      current = next
-      ;(microsteps ??= []).push(step)
-      if (transitionResult.commands.length > 0) {
-        ;(commands ??= []).push(...transitionResult.commands)
-      }
-      if (transitionResult.raisedEvents.length > 0) {
-        ;(raisedEvents ??= []).push(...transitionResult.raisedEvents)
-      }
-      if (transitionResult.emittedEvents.length > 0) {
-        ;(emittedEvents ??= []).push(...transitionResult.emittedEvents)
-      }
-    }
-
-    if (descriptor.nodes[current.activeLeaves[0]!]!.type === "final") {
-      continue
-    }
-    const raised = raisedEvents?.[raisedIndex]
-    if (raised === undefined) {
-      return {
-        next: current,
-        commands: commands ?? emptyCompiledValues,
-        emittedEvents: emittedEvents ?? emptyCompiledValues,
-        microsteps: microsteps ?? emptyCompiledValues,
-        done: false,
-        output: undefined
-      }
-    }
-    raisedIndex += 1
-    event = raised
-  }
-}
-
-const planIndexedConfiguration = (
-  machine: Machine.Any,
-  descriptor: IndexedExecutionDescriptor,
-  configuration: IndexedConfiguration,
-  input: unknown
-): MacrostepPlan<IndexedConfiguration, any, any, any, any> => {
-  const decoded = decodeEventSync(machine, input) as { readonly _tag: PropertyKey }
-  if (descriptor.flat) {
-    return planIndexedFlatConfiguration(machine, descriptor, configuration, decoded)
-  }
-  if (descriptor.finalIndices.some((index) => configuration.active[index] === 1)) {
-    const active = activeConfigurationFromIndexed(descriptor, configuration)
-    if (isActiveFinalConfiguration(machine, active)) {
-      const completed = completeConfigurationSync(machine, active, decoded).configuration
-      const root = getRootPath(machine, completed)
-      if (!completed.outputs.has(root)) {
-        throw new Error("Machine reached a terminal indexed configuration without a completed root output")
-      }
-      return {
-        next: indexedConfigurationFromActive(descriptor, completed),
-        commands: [],
-        emittedEvents: [],
-        microsteps: [],
-        done: true,
-        output: completed.outputs.get(root)
-      }
-    }
-  }
-
-  const selections = selectIndexedEventTransitions(machine, descriptor, configuration, decoded)
-  if (selections.length === 0) {
-    return {
-      next: configuration,
-      commands: [],
-      emittedEvents: [],
-      microsteps: [],
-      done: false,
-      output: undefined
-    }
-  }
-
-  const first = indexedMicrostep(machine, descriptor, configuration, decoded, selections)
-  let current = first.next
-  let currentEvent: any = decoded
-  const commands = [...first.commands]
-  const raisedEvents = [...first.raisedEvents]
-  const emittedEvents = [...first.emittedEvents]
-  const microsteps = [first]
-  let raisedIndex = 0
-  let iterations = 0
-
-  while (true) {
-    iterations += 1
-    if (iterations > MaxMacrostepIterations) {
-      throw new InfiniteTransitionError({
-        machineId: machine.id,
-        state: descriptor.nodes[descriptor.leafIndices.find((index) => current.active[index] === 1)!]!.path,
-        maxIterations: MaxMacrostepIterations
-      })
-    }
-
-    if (descriptor.finalIndices.some((index) => current.active[index] === 1)) {
-      const completed = completeConfigurationSync(
-        machine,
-        activeConfigurationFromIndexed(descriptor, current),
-        currentEvent
-      ).configuration
-      current = indexedConfigurationFromActive(descriptor, completed)
-      if (isActiveFinalConfiguration(machine, completed)) {
-        const root = getRootPath(machine, completed)
-        if (!completed.outputs.has(root)) {
-          throw new Error("Machine reached a terminal indexed configuration without a completed root output")
-        }
-        return {
-          next: current,
-          commands,
-          emittedEvents,
-          microsteps,
-          done: true,
-          output: completed.outputs.get(root)
-        }
-      }
-    }
-
-    const raised = raisedEvents[raisedIndex]
-    if (raised === undefined) {
-      return {
-        next: current,
-        commands,
-        emittedEvents,
-        microsteps,
-        done: false,
-        output: undefined
-      }
-    }
-    raisedIndex += 1
-    currentEvent = raised
-    const raisedSelections = selectIndexedEventTransitions(machine, descriptor, current, raised)
-    if (raisedSelections.length === 0) continue
-    const step = indexedMicrostep(machine, descriptor, current, raised, raisedSelections)
-    current = step.next
-    commands.push(...step.commands)
-    raisedEvents.push(...step.raisedEvents)
-    emittedEvents.push(...step.emittedEvents)
-    microsteps.push(step)
-  }
-}
-
-export interface CompiledExecutionPlan {
-  readonly fromConfiguration: (configuration: ActiveConfiguration) => unknown
-  readonly toConfiguration: (state: unknown) => ActiveConfiguration
-  readonly snapshot: (state: unknown) => Machine.Snapshot<any>
-  readonly plan: (
-    state: unknown,
-    event: unknown
-  ) => MacrostepPlan<any, any, any, any, any>
-  readonly initial?: (
-    args: ReadonlyArray<unknown>
-  ) => {
-    readonly state: Machine.Snapshot<any>
-    readonly configuration: unknown
-    readonly activeConfiguration: ActiveConfiguration
-    readonly initialEntryPaths: ReadonlyArray<string>
-    readonly done: boolean
-    readonly output: unknown
-  }
-}
-
-const executionPlanCache = new WeakMap<Machine.Any, CompiledExecutionPlan>()
-
-const makeActiveExecutionPlan = (machine: Machine.Any): CompiledExecutionPlan => ({
-  fromConfiguration: (configuration) => configuration,
-  toConfiguration: (state) => state as ActiveConfiguration,
-  snapshot: (state) => snapshotFromConfiguration(machine, state as ActiveConfiguration),
-  plan: (state, event) => planConfiguration(machine as any, state as ActiveConfiguration, event as any)
-})
-
-const makeIndexedExecutionPlan = (
-  machine: Machine.Any,
-  indexed: IndexedExecutionDescriptor
-): CompiledExecutionPlan => ({
-  fromConfiguration: (configuration) => indexedConfigurationFromActive(indexed, configuration),
-  toConfiguration: (state) => activeConfigurationFromIndexed(indexed, state as IndexedConfiguration),
-  snapshot: (state) => snapshotFromIndexed(indexed, state as IndexedConfiguration),
-  plan: (state, event) => planIndexedConfiguration(machine, indexed, state as IndexedConfiguration, event),
-  initial: (args) => {
-    const inputArgs = machine.input === undefined
-      ? args
-      : args.length === 0
-      ? (decodeInputSync(machine, machine.input, undefined), args)
-      : [decodeInputSync(machine, machine.input, args[0])]
-    const initial = machine.initial(...inputArgs as any)
-    const active = normalizeConfigurationSync(machine, initial as Machine.Snapshot<any>)
-    validateInitialConfiguration(machine, active)
-    const completed = completeConfigurationSync(machine, active, InitialEvent).configuration
-    const configuration = indexedConfigurationFromActive(indexed, completed)
-    const state = snapshotFromIndexed(indexed, configuration)
-    const done = isActiveFinalConfiguration(machine, completed)
-    if (!done) {
-      return {
-        state,
-        configuration,
-        activeConfiguration: completed,
-        initialEntryPaths: getInitialEntryPaths(machine, completed),
-        done: false,
-        output: undefined
-      }
-    }
-    const root = getRootPath(machine, completed)
-    if (!completed.outputs.has(root)) {
-      throw new Error("Machine reached a terminal configuration without a completed root output")
-    }
-    return {
-      state,
-      configuration,
-      activeConfiguration: completed,
-      initialEntryPaths: getInitialEntryPaths(machine, completed),
-      done: true,
-      output: completed.outputs.get(root)
-    }
-  }
-})
-
-export type ExecutionPlanStrategy = "generic" | "indexed-flat" | "indexed-hierarchical" | "auto"
-
-export interface SelectedExecutionPlan {
-  readonly strategy: Exclude<ExecutionPlanStrategy, "auto">
-  readonly plan: CompiledExecutionPlan
-}
-
-const selectExecutionPlan = (
-  machine: Machine.Any,
-  strategy: ExecutionPlanStrategy
-): SelectedExecutionPlan => {
-  if (strategy === "generic") {
-    return { strategy, plan: makeActiveExecutionPlan(machine) }
-  }
-  const indexed = compileIndexedExecutionDescriptor(machine)
-  if (indexed === undefined) {
-    if (strategy === "auto") {
-      return { strategy: "generic", plan: makeActiveExecutionPlan(machine) }
-    }
-    throw new Error(`Machine cannot compile the requested ${strategy} execution plan`)
-  }
-  const selected = indexed.flat ? "indexed-flat" : "indexed-hierarchical"
-  if (strategy !== "auto" && strategy !== selected) {
-    throw new Error(`Machine compiled ${selected}, not the requested ${strategy} execution plan`)
-  }
-  return { strategy: selected, plan: makeIndexedExecutionPlan(machine, indexed) }
-}
-
-/** @internal Test-only uncached strategy selection. */
-export const selectExecutionPlanForTesting = (
-  machine: Machine.Any,
-  strategy: ExecutionPlanStrategy
-): SelectedExecutionPlan => selectExecutionPlan(machine, strategy)
-
-export const compileExecutionPlan = (machine: Machine.Any): CompiledExecutionPlan => {
-  const cached = executionPlanCache.get(machine)
-  if (cached !== undefined) {
-    return cached
-  }
-  const indexed = compileIndexedExecutionDescriptor(machine)
-  const compiled = indexed === undefined
-    ? makeActiveExecutionPlan(machine)
-    : makeIndexedExecutionPlan(machine, indexed)
-  executionPlanCache.set(machine, compiled)
-  return compiled
 }
 
 const snapshotMacrostep = <
